MonthlyRatioAutomation/requirements.txt
camelot_py==0.8.0
SQLAlchemy==1.3.17
requests==2.23.0
pyjarowinkler==1.8
beautifulsoup4==4.9.1
camelot==13.04.13-gpl-pyqt
python_dateutil==2.8.1
